Move paginate to package; Add Query to allow composition on All
